ready for turn in

+++ application/Main.java
@@ -23,7 +23,7 @@ public class Main extends Application {
   // declares screen length, width and padding around application.
   protected static final int SCREEN_LENGTH = 1200;
   protected static final int SCREEN_WIDTH = 900;
-  protected static final int SCREEN_PADDING = 10;
+  protected static final double SCREEN_PADDING = 10;
 
   /**
    * Sets up the primary stage.
